refactor(auth): streamline authentication forms and schemas, removing unused fields and consolidating validation logic
